refactor: extract State into js/state.js
